Revert "Props usestate"
