+++ src/App.jsx
@@ -1,20 +1,24 @@
-import { useState } from 'react';
-import TodoList from './TodoList';
-import AddTodoForm from './AddTodoForm';
- 
-function App() {
 
-  const [newTodo,setNewTodo] = useState('');
 
-  return (
+const todoList = [
+  { id: 1, title: "Drink a glass of water" },
+  { id: 2, title: "Study for upcoming Python quiz" },
+  { id: 3, title: "Create R&D strategy for wellness project" }
+];
+
+function App (){
+  return(
     <div>
-      <h1>Add a list Item</h1>
-      <AddTodoForm onAddTodo={setNewTodo} />
-      <p>New Todo: {newTodo}</p>
-      <TodoList />
+      <h1>
+        Daily To-do List
+      </h1>
+      <ul>
+        {todoList.map((item) => (
+          <li key={item.id}>{item.title}</li>
+        ))}
+      </ul>
     </div>
-
   );
-}
+};
 
 export default App;
